Un-nest class
Move the IHttpMessageHandlerBuilderFilter in the samples into its own class so it reads better in the README snippet.
